Split provider IPC handlers
